==> 04_eqtm/04_analyse_eqtm_mad_25.Rmd <==
---
title: "eQTM_analysis"
author: "Anastasiia"
date: "4/24/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r load-libs}
library(data.table)
library(dplyr)
library(ggplot2)
```

# eQTM analysis

```{r setup-param-and-load-data}
treatment <- "dex"

eqtm.pre     <- "~/bio/datasets/eQTM/"
eqtm.res.pre <- paste0(eqtm.pre, "result/")

methyl.layer.fn <- paste0(eqtm.pre, "methyl_beta_mtrx_", treatment, "_beta_0_p_90.csv")
methyl.layer    <- fread(methyl.layer.fn)

cis.eqtm.veh.fn <- paste0(eqtm.res.pre, "eqtm_cis_result_", treatment, ".csv")
cis.eqtm.veh    <- fread(cis.eqtm.veh.fn)
```

```{r}
colnames(cis.eqtm.veh)

cis.eqtm <- cis.eqtm.veh[beta != 0, ] 

paste("Number of unique CpGs:", length(unique(cis.eqtm$SNP)))
paste("Number of unique genes:", length(unique(cis.eqtm$gene))) 
```

```{r unique-cpgs-plot, cache = F}
# prior.cpg.ids  <- as.data.frame(cbind("CpG_ID" = unique(prior.cpg.gex$CpG_ID), "Type" = "Prior"))
prior.cpg.ids <- as.data.frame(cbind("CpG_ID" = unique(rownames(methyl.layer)), "Type" = "Prior"))
eqtm.cpg.ids  <- as.data.frame(cbind("CpG_ID" = unique(cis.eqtm$SNP), "Type" = "cis-eQTM"))

cpgs.analysis <- rbind(prior.cpg.ids, eqtm.cpg.ids)

ggplot(cpgs.analysis, aes(Type)) + 
  geom_bar(stat = "count", fill = "purple") +
  coord_flip() +
  geom_text(stat = "count", aes(label=..count..), 
     hjust = 0.5, color = "black", position = position_dodge(0.9), size = 3.5) +
   labs(title = "Nr of unique CpGs before filtering")
```

```{r}
fdr.thrsh <- 0.01
beta.trsh <- 0
cis.eqtm <- cis.eqtm.veh[FDR < fdr.thrsh & abs(beta) > beta.trsh] %>% unique()

cis.eqtm$SNP %>% unique() %>% length()

eqtm.cpg.ids.fdr.001  <- as.data.frame(cbind("CpG_ID" = unique(cis.eqtm$SNP), "Type" = "cis-eQTM_FDR_001"))
cpgs.analysis <- rbind(prior.cpg.ids, eqtm.cpg.ids, eqtm.cpg.ids.fdr.001)

ggplot(cpgs.analysis, aes(Type)) + 
  geom_bar(stat = "count", fill = "purple") +
  coord_flip() +
  geom_text(stat = "count", aes(label=..count..), 
     hjust = 0.5, color = "black", position = position_dodge(0.9), size = 3.5) +
   labs(title = "Nr of unique CpGs")

# head(cis.eqtm[order(FDR, decreasing = F)])
# nrow(cis.eqtm[,.(SNP, gene)] %>% unique())
```
# KiMONo analysis
```{r}
data.kimono.pre    <- "~/bio/datasets/kimono/"
map.cpg.tbl.fn  <- paste0(data.kimono.pre, "/mapping/mapping_cpg_gene_ensg_full.csv")
map.cpg.gene.ensg.tbl <- read.csv2(map.cpg.tbl.fn)

cis.eqtm.gene <- inner_join(cis.eqtm, map.cpg.gene.ensg.tbl, by = c("gene" = "Ensemble_ID", "SNP" = "CpG_ID")) # %>% dplyr::select(CpG = SNP, ENSG_ID = gene, beta, t-stat, p-value, FDR)

head(cis.eqtm.gene[order(FDR, decreasing = F)])
```

```{r}
kimono.res.fn         <- paste0(data.kimono.pre, "output/kimono_res_dex.csv") # experiment nr 5
kimono.res            <- fread(kimono.res.fn)
kimono.net <- kimono.res[value != 0, ] 
kimono.methyl.expr <- kimono.net[relation == "methyl_expr"] %>% unique()
kimono.methyl.gene <- inner_join(kimono.methyl.expr, map.cpg.gene.ensg.tbl, by = c("target" = "Ensemble_ID", "predictor" = "CpG_ID"))

rsq.trsh  <- 0.01

kimono.methyl.gene <- kimono.methyl.gene %>% # filter((value > beta.trsh) | (value < (-beta.trsh))) %>%
  filter(r_squared > rsq.trsh) %>%
  filter(predictor != '(Intercept)') %>% setDT

head(kimono.methyl.gene[order(r_squared, decreasing = F)])
```
```{r}
cis.eqtm$SNP %>% unique %>% length
monigenes <- unique(kimono.methyl.gene$target); length(monigenes) # genes that are found in moni
myeqtl <- cis.eqtm[ gene %in% monigenes,] # find them in eqtl
coef2 <- kimono.methyl.gene[ target %in% myeqtl$gene,] # 
all(as.character(unique(myeqtl$gene)) %in% as.character(unique(coef2$target)))
paste("further analysis with", length(unique(coef2$target)), "genes that were present in both eQTL and MONI")

```


```{r venndiagram-eqtm-kimono}
library(ggVennDiagram)
ggVennDiagram(list(cis_eqtm = cis.eqtm.gene$Gene_ID, kimono = kimono.methyl.gene$Gene_ID), scaled = T) + theme(legend.position = "none")
```
